+++ app-updater/src/main/java/com/king/app/updater/notification/NotificationHandler.kt
@@ -46,8 +46,8 @@ open class NotificationHandler : INotificationHandler {
         smallIcon: Int,
         title: CharSequence,
         content: CharSequence,
-        progress: Int,
-        total: Int,
+        progress: Long,
+        total: Long,
         supportCancelDownload: Boolean
     ) {
         NotificationUtils.showProgressNotification(
